server/requirements.txt
Flask[async]
Flask-SQLAlchemy
Flask-Caching
SQLAlchemy
psycopg2-binary
python-dotenv
aiohttp
asyncio
Gunicorn